small refactor and add sliding window test
