+++ src/test/test_utils.py
@@ -12,6 +12,7 @@ from src.utils import (
     filter_blocks,
     get_ids_with_suffix,
     encode_parser_output,
+    sliding_window,
 )
 
 
@@ -155,3 +156,15 @@ def test_encode_indexer_input(test_pdf_file_json):  # noqa: F811
 
 # TODO get_Text2EmbeddingsInput_array
 #   TODO needs s3 files, local files, of the form json IndexerInput objects
+
+
+def test_sliding_window():
+    """Tests that the sliding_window function returns the correct embeddings."""
+    text = "Hello world! " * 50
+    window_size = 10
+    stride = 5
+
+    windows = sliding_window(text=text, window_size=window_size, stride=stride)
+
+    assert windows[0] == "Hello worl"
+    assert windows[1] == " world! He"
